+++ source/public/Enable-ChocolateySource.ps1
@@ -40,7 +40,7 @@ function Enable-ChocolateySource
 
     process
     {
-        if (-not ($chocoCmd = Get-Command 'choco.exe' -CommandType Application -ErrorAction SilentlyContinue))
+        if (-not ($chocoCmd = @(Get-Command 'choco.exe' -CommandType 'Application' -ErrorAction 'SilentlyContinue')[0]))
         {
             throw "Chocolatey Software not found."
         }