+++ source/public/Get-ChocolateyPackage.ps1
@@ -119,7 +119,7 @@ function Get-ChocolateyPackage
 
     process
     {
-        if (-not ($chocoCmd = Get-Command 'choco.exe' -CommandType 'Application' -ErrorAction 'SilentlyContinue'))
+        if (-not ($chocoCmd = @(Get-Command 'choco.exe' -CommandType 'Application' -ErrorAction 'SilentlyContinue')[0]))
         {
             throw "Chocolatey Software not found."
         }