+++ source/public/Get-ChocolateySetting.ps1
@@ -32,7 +32,7 @@ function Get-ChocolateySetting
     )
     begin
     {
-        if (-not ($chocoCmd = Get-Command 'choco.exe' -CommandType Application -ErrorAction SilentlyContinue))
+        if (-not ($chocoCmd = @(Get-Command 'choco.exe' -CommandType 'Application' -ErrorAction 'SilentlyContinue')[0]))
         {
             throw "Chocolatey Software not found."
         }